Improve: iptables-nft and iptables

+++ install.sh
@@ -9,6 +9,7 @@ function modules_chk()
         nft --version > /dev/null 2>&1
         if [ $? -eq 0 ]; then
              local extra_module="xtables-nft iptables-nft ip6tables-nft kmod-ipt-quota2"
+             export nft=1;
         fi
 
         opkg install $extra_module iptables-mod-quota2 > /dev/null 2>&1      
@@ -36,14 +37,16 @@ function install()
     echo '1 0 * * * /etc/quota.d/reset_bytescounter' >> /etc/crontabs/root
     echo >> /etc/crontabs/root
 
-    echo '/etc/quota.d/restore_bytescounter' > /etc/firewall.quota
+    if [ -z $nft ]; then
+        echo '/etc/quota.d/restore_bytescounter' > /etc/firewall.quota
 
-    uci -q delete firewall.quota
-    uci set firewall.quota="include"
-    uci set firewall.quota.path="/etc/firewall.quota"
-    uci set firewall.quota.reload="1"
-    uci commit firewall
-    
+        uci -q delete firewall.quota
+        uci set firewall.quota="include"
+        uci set firewall.quota.path="/etc/firewall.quota"
+        uci set firewall.quota.reload="1"
+        uci commit firewall
+    fi
+
     /etc/init.d/firewall restart > /dev/null 2>&1
     /etc/init.d/cron restart > /dev/null 2>&1
 
@@ -54,11 +57,15 @@ function uninstall()
 {
     rm -r /etc/quota.d
     rm /etc/vars
-    rm /etc/firewall.quota
 
-    uci -q delete firewall.quota
-    uci commit firewall
-    
+    nft --version > /dev/null 2>&1
+    if [ $? -eq 1 ]; then 
+        rm /etc/firewall.quota
+
+        uci -q delete firewall.quota
+        uci commit firewall
+    fi
+
     /etc/init.d/firewall restart > /dev/null 2>&1
     /etc/init.d/cron restart > /dev/null 2>&1
